pyproject.toml
[tool.poetry]
name = "spread-wings-bot"
version = "0.1.0"
description = "add later"
authors = ["add later"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11.1"
python-telegram-bot = "^20.3"
pydantic = {extras = ["email"], version = "^1.10.8"}
python-dotenv = "^1.0.0"
requests-cache = "^1.0.1"
attrs = "^23.1.0"
redis = "^4.5.5"
django = "^4.2.3"
uvicorn = "^0.23.0"
django-asgi-lifespan = "^0.1.0"
psycopg2-binary = "^2.9.6"
django-environ = "^0.10.0"
async-lru = "^2.0.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
pre-commit = "^3.3.1"
black = "^23.3.0"
pytest = "^7.3.1"
anyio = "^3.7.0"
pytest-tornasync = "^0.6.0.post2"
factory-boy = "^3.2.1"


[tool.pytest.ini_options]
addopts = "-vv -p no:cacheprovider -p no:warnings"
testpaths = [
    "tests",
]
python_files = "test_*.py"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
